seeded both dbs with 10mill items

+++ db/PostgresDB/itemSchema.sql
@@ -1,7 +1,7 @@
 CREATE TABLE items (
   id SERIAL,
   itemname VARCHAR(255),
-  price INT,
+  price NUMERIC,
   description1 TEXT,
   description2 TEXT
 );
